Refactor move multigascollector to arbos
