+++ arbos/multigascollector/factory.go
