+++ arbos/multigascollector/file_collector_test.go
@@ -17,7 +17,7 @@ import (
 
 	"github.com/ethereum/go-ethereum/arbitrum/multigas"
 
-	"github.com/offchainlabs/nitro/execution/multigascollector/proto"
+	"github.com/offchainlabs/nitro/arbos/multigascollector/proto"
 )
 
 func TestIdleCollector(t *testing.T) {
