+++ arbos/multigascollector/interface.go
